Embed longest sentences first to tune batch size
Just a suggestion, but it might help to process the longest texts first, in order to catch OOMs at the beginning of BERT-scoring, instead of at the end. This would help to quickly tune batch size to the largest possible value given your current hardware.

The code is otherwise pretty optimized!